Changed new Field version numbers to all be int's (some were doubles) Moved gradient code into the Fields

+++ src/Core/Datatypes/Field.cc
@@ -16,7 +16,7 @@ Field::~Field()
 {
 }
 
-const double FIELD_VERSION = 1.0;
+const int FIELD_VERSION = 1;
 
 void 
 Field::io(Piostream& stream){
